refactor(logic): eliminate world→diplomacy DAG edge for #3290 Phase 0
Hoist faction membership and read-only relation lookups into world/
modules; diplomacy re-exports preserve existing import paths.
