+++ packages/colonizethis_logic/lib/src/world/naval_resolution.dart
@@ -4,7 +4,7 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 
 import '../combat/naval_combat_resolver.dart';
 import '../constants.dart';
-import '../diplomacy/diplomacy_relation_lookup.dart';
+import 'diplomatic_relation_lookup.dart';
 import '../dossier/evidence_rules.dart';
 import '../dossier/event_dialogue.dart';
 import '../event_bus/game_event_bus.dart';
