+++ tool/check_logic_domain_import_dag.dart
@@ -52,12 +52,6 @@ const _grandfatherAllowlist = <String>{
   'orders->turn:orders/order_projections.dart',
   'orders->turn:orders/orders_application.dart',
   'orders->turn:orders/orders_application_context.dart',
-  'world->diplomacy:world/civilian_ownership_legality.dart',
-  'world->diplomacy:world/civilian_tile_occupancy.dart',
-  'world->diplomacy:world/connectivity_resolver.dart',
-  'world->diplomacy:world/naval_mission_orders.dart',
-  'world->diplomacy:world/naval_resolution.dart',
-  'world->diplomacy:world/province_ownership_transfer.dart',
   'combat->diplomacy:combat/military_attack_economy.dart',
   'combat->diplomacy:combat/naval_combat_resolver.dart',
   'combat->diplomacy:combat/unopposed_province_capture.dart',
